core: reduce allowed stack depth
Otherwise we hit stack overflow.

In the future we need to change how we process function non-native
function calls in the VM, we need to go deeper in the VM stack without
calling back to exec_function().  This will reduce native stack usage
and also it may help with the implementation of tail recursion.